fix(patina): emit lsp diagnostics as one json array
